Refactor bot handlers into modules
