+++ admin/app.py
@@ -10,6 +10,7 @@ from .routers import auth as auth_router
 from .routers import configs as config_router
 from .routers import servers as server_router
 from .routers import users as user_router
+from .dependencies import parse
 
 app = FastAPI()
